feat: add and refactor post-merge proofs
