+++ ethportal-api/src/types/consensus/mod.rs
@@ -4,7 +4,7 @@ pub mod body;
 pub mod execution_payload;
 pub mod fork;
 pub mod header;
-pub mod header_proof;
+pub mod historical_summaries;
 pub mod kzg_commitment;
 pub mod light_client;
 pub mod participation_flags;
