+++ ethportal-api/src/types/execution/mod.rs
@@ -1,5 +1,6 @@
 pub mod accumulator;
 pub mod block_body;
 pub mod header;
+pub mod header_with_proof;
 pub mod receipts;
 pub mod transaction;
